backend/python/requirements.txt
pymupdf==1.23.8
Pillow==10.1.0
